generalized link button helper for all sites

+++ app/helpers/application_helper.rb
@@ -1,11 +1,13 @@
 module ApplicationHelper
 
-	def admin_links(object, edit_path)
-		(link_to 'Zurück', url_for(:back), class:'btn btn-sm btn-warning') + ' ' +
-		(link_to 'Bearbeiten', edit_path, class: "btn btn-sm btn-warning") + ' ' +
-		(link_to 'Löschen', object,
-												method: :delete,
-												data: { confirm: 'Are you sure?' }, class: "btn btn-sm btn-warning")
+	def link_button_helper(object = "", edit_path = "")
+		(link_to 'Zurück', url_for(:back), class: "btn btn-warning") + ' ' +
+		unless object == ""
+			(link_to 'Bearbeiten', edit_path, class: "btn btn-warning") + ' ' +
+			(link_to 'Löschen', object,
+													method: :delete,
+													data: { confirm: 'Are you sure?' }, class: "btn btn-warning")
+		end
 	end
 
   def active?(key, value)
